+++ src/Data/Time/Calendar/BankHoliday/Germany/ExtraHolidays.hs
@@ -69,7 +69,7 @@ only includes some states.
  - Rheinland-Pfalz: https://mdi.rlp.de/themen/buerger-und-staat/verfassung-und-verwaltung/sonn-und-feiertagsrecht
  - Brandenburg: https://bravors.brandenburg.de/gesetze/ftg_2003/6
  - Sachsen-Anhalt: https://www.landesrecht.sachsen-anhalt.de/bsst/document/jlr-FeiertGSTrahmen/part/X
- - Thüringen: https://buerger.thueringen.de/detail?pstId=354718
+ - Thüringen: https://landesrecht.thueringen.de/bsth/document/jlr-FeiertGTHrahmen
  - Hamburg: https://www.hamburg.de/ferien-und-feiertage/
  - Mecklenburg-Vorpommern u. Berlin (Frauentag): https://www.deutsche-rentenversicherung.de/DRV/DE/Ueber-uns-und-Presse/Presse/Meldungen/2024/240306_frauentag_feiertag_frei.html
  - Saarland: https://www.saarland.de/mibs/DE/themen-aufgaben/aufgaben/buerger_und_staat/sonn_u_feiertagsrecht/feiertagsrecht_node.html
@@ -124,6 +124,7 @@ data ExtraHoliday
   | MariaeHimmelfahrt      -- ^ Mariä Himmelfahrt (Bayern*, Saarland)
   | Allerheiligen          -- ^ Allerheiligen (Bayern, Baden-Württemberg, Nordrhein-Westfalen, Rheinland-Pfalz, Saarland)
   | Reformationstag        -- ^ Reformationstag (Niedersachsen, Sachsen, Schleswig-Holstein, Brandenburg, Sachsen-Anhalt, Thüringen, Hamburg, Mecklenburg-Vorpommern, Bremen)
+  | Weltkindertag          -- ^ Weltkindertag
   | InternationalerFrauentag -- ^ Internationaler Frauentag (Berlin, Mecklenburg-Vorpommern)
   | BussUndBettag          -- ^ Buß- und Bettag (Sachsen)
   deriving (Enum, Eq, Bounded, Show, Read)
@@ -140,6 +141,7 @@ toDay year MariaeHimmelfahrt       = fromGregorian year 8 15
 toDay year Allerheiligen           = fromGregorian year 11 1
 toDay year InternationalerFrauentag = fromGregorian year 3 8
 toDay year Reformationstag          = fromGregorian year 10 31
+toDay year Weltkindertag           = fromGregorian year 9 20
 toDay year BussUndBettag           = calculateBussUndBettag year
 
 -- | Compute 'Maybe' the holiday for a given date.
@@ -171,7 +173,8 @@ germanHolidayName d = case d of
   Friedensfest           -> "Friedensfest"
   MariaeHimmelfahrt      -> "Mariä Himmelfahrt"
   Allerheiligen          -> "Allerheiligen"
-  Reformationstag          -> "Reformationstag"
+  Reformationstag        -> "Reformationstag"
+  Weltkindertag          -> "Weltkindertag"
   InternationalerFrauentag -> "Internationaler Frauentag"
   BussUndBettag            -> "Buß- und Bettag"
 
@@ -211,6 +214,7 @@ isHolidayInState SachsenAnhalt HeiligeDreiKoenige = True
 isHolidayInState SachsenAnhalt Reformationstag = True
 isHolidayInState Thueringen Fronleichnam = True
 isHolidayInState Thueringen Reformationstag = True
+isHolidayInState Thueringen Weltkindertag = True
 isHolidayInState Hamburg Reformationstag = True
 isHolidayInState MecklenburgVorpommern Reformationstag = True
 isHolidayInState MecklenburgVorpommern InternationalerFrauentag = True